pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crumbbum"
version = "0.1.0"
description = "Take control of your browser cookies via Chrome DevTools Protocol"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
  { name = "CrumbBum Maintainers" }
]
dependencies = [
  "requests>=2.28",
  "websocket-client>=1.6",
]

[project.scripts]
crumbbum = "crumbbum.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.format]
quote-style = "preserve"

[tool.mypy]
python_version = "3.8"
strict = false
ignore_missing_imports = true
